refactor(duplicati-r2-list): extract shared resolver into duplicati_r2_common
Cut B of #204 needs the same slug -> SQLite resolution, snapshot lookup,
path normalization, exit-code conventions, and table emitter that Cut A
ships. Move them out of `duplicati_r2_list.py` into a sibling
`duplicati_r2_common.py` that both binaries import, so future
operator-facing tools cannot drift from the resolver Cut A validated.

`list.nix` now installs both `.py` files into
`$out/libexec/duplicati-r2-tools/` and symlinks
`$out/bin/duplicati-r2-list` at the entry-point script; the
`sys.path.insert(0, dirname(realpath(__file__)))` shim resolves the
sibling import regardless of how the script is launched (direct path,
$out/bin symlink, $PATH lookup).

Cut A behaviour is unchanged: the existing `installCheckPhase` still
exercises `versions` / `ls` / `stat` / `history` / `grep`, JSON output,
exit codes, and the manifest fallback path.
